error fix

+++ schema/1_init.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS metrics (
     id TEXT PRIMARY KEY,
     type TEXT NOT NULL,
-    delta INT,
+    delta BIGINT,
     value DOUBLE PRECISION
 );
